Restore Font Awesome shortcode
